refactor(repository): refactored v1 encryption overhead to be a function that's only invoked when actual V1 index is opened (#2300)

+++ repo/content/committed_read_manager.go
@@ -472,7 +472,7 @@ func (sm *SharedManager) setupReadManagerCaches(ctx context.Context, caching *Ca
 	sm.metadataCache = metadataCache
 	sm.indexBlobCache = indexBlobCache
 	sm.committedContents = newCommittedContentIndex(caching,
-		uint32(sm.format.Encryptor().Overhead()),
+		sm.format.Encryptor().Overhead,
 		sm.format,
 		sm.enc.getEncryptedBlob,
 		sm.namedLogger("committed-content-index"),
